Make the constructor that create Lifetime with AnyObject public
